+++ packages/core/modules/i18n/validation/constains.js
@@ -1,0 +1,10 @@
+// Note: Don't use ":" char inside strings, it's reserved as a namespace
+
+export const EMPTY_GROUP = "EMPTY_GROUP";
+export const NO_CONFIG_FOR_FIELD = "NO_CONFIG_FOR_FIELD";
+export const NO_CONFIG_FOR_OPERATOR = "NO_CONFIG_FOR_OPERATOR";
+export const FIXED_OPERATOR = "FIXED_OPERATOR";
+export const NO_CONFIG_FOR_FUNCTION = "NO_CONFIG_FOR_FUNCTION";
+export const INVALID_VALUE = "INVALID_VALUE";
+export const INVALID_FUNC_ARG_VALUE = "INVALID_FUNC_ARG_VALUE";
+